Integrated LS1 and LS2 into the timing framework

+++ code/tsp_ls1.py
@@ -7,7 +7,9 @@ import random
 import numpy as np
 import math
 import time
-from multiprocessing import Process, Pipe, Queue
+import copy
+
+from multiprocessing import Process, Pipe
 
 
 def distance_calculator(nodes):
@@ -32,7 +34,7 @@ def get_min_cost(dest_group,rem_nodes):
 #This will solve a greedy problem to give us a place to start
 def greedy_heuristic(dist_matrix,index_start):
     initialdist_matrix = dist_matrix
-    start_index = index_start#random.randint(0,dist_matrix.shape[0]-1)
+    start_index = index_start #random.randint(0,dist_matrix.shape[0]-1)
     total_cost = 0
     route = []
     rem_nodes = np.ones(dist_matrix.shape[0])
@@ -56,29 +58,36 @@ def calculate_route_cost(route,cost_matrix):
 
 
 def two_opt(dist_matrix,nodes,start_nodes,tracepoint_pipe : Pipe, solution_pipe : Pipe):
-    initial = 1
     start_time = time.process_time()
+    initial = 1
     mat = distance_calculator(nodes)
-    while 1:
-        for x in start_nodes:
-            greedy_route,greedy_total_cost = greedy_heuristic(mat,x)
-            route_matrix = greedy_route
-            if initial == 1:
-                bestroute = greedy_route
-                bestcost = greedy_total_cost
-                initial = 0
-            for i in range(1,len(route_matrix)-1):
-                for k in range(i+i,len(route_matrix)):
-                    improved_route = route_matrix[:]
-                    improved_route[i:k] = route_matrix[k-1:i-1:-1] #Here we will take the section to flip and flip it and replace the order.
-                    new_cost = calculate_route_cost(improved_route,dist_matrix)
-                    if new_cost < bestcost:
-                        print("Improvement by 2opt made")
-                        bestroute = improved_route
-                        bestcost = new_cost
-                        solution_pipe.send(Solution( bestcost, bestroute))
-                        tracepoint_pipe.send( Tracepoint( time.process_time() - start_time, bestcost ) )
-            route_matrix = bestroute
+
+    for x in start_nodes:
+        greedy_route,greedy_total_cost = greedy_heuristic(mat,x)
+        route_matrix = greedy_route # Use greedy heuristic solution as the starting point
+        if initial == 1:
+            bestroute = greedy_route
+            bestcost = greedy_total_cost
+            initial = 0
+        for i in range(1,len(route_matrix)-1):
+            for k in range(i+i,len(route_matrix)):
+                improved_route = route_matrix[:]
+                improved_route[i:k] = route_matrix[k-1:i-1:-1] #Here we will take the section to flip and flip it and replace the order.
+                new_cost = calculate_route_cost(improved_route,dist_matrix)
+                if new_cost < bestcost:
+                    # print("Improvement by 2opt made")
+                    bestroute = improved_route
+                    bestcost = new_cost
+
+                    # Construct a solution and send it to output pipe
+                    solution = Solution( bestcost )
+                    for node in bestroute:
+                        solution.node_list.append( copy.deepcopy(node) )
+                    solution_pipe.send( solution )
+                    tracepoint_pipe.send( Tracepoint( time.process_time() - start_time, bestcost ) )
+        route_matrix = bestroute
+
+    exit( 0 ) # Exit process since all possible greedy startpoints have been hill-climbed.
 
 def printmat(mat):
     for i in range(0,len(mat)):
@@ -87,7 +96,7 @@ def printmat(mat):
         print()    
 
 def ls1( nodes , timeout : int,seed_num ):
-    start_time = time.time()
+    start_time = time.process_time()
     solution_read, solution_write = Pipe()
     tracepoint_read, tracepoint_write = Pipe()
     random.seed(seed_num)
@@ -100,22 +109,19 @@ def ls1( nodes , timeout : int,seed_num ):
 
     p = Process( target = two_opt, args = (mat,nodes,start_nodes,tracepoint_write, solution_write ) ) 
     p.start() # Start the process
-    while( time.time()-start_time < timeout ):
-        if solution_read.poll(timeout - ( time.time() - start_time )) and\
-                tracepoint_read.poll(timeout - ( time.time() - start_time )) :
+
+    # Have the thread spin so that it keeps track of time most accurately
+    while( ( time.process_time()-start_time ) < timeout and p.is_alive() ):
+        if solution_read.poll( 0 ) and tracepoint_read.poll( 0 ):
             # Block until a solution is available or timeout
             solution = solution_read.recv( )
-            
+
             # Receive corresponding tracepoint and append
-            tracepoint = tracepoint_read.recv()
-            trace.add_tracepoint(float(tracepoint.time),tracepoint.quality)
-            print("New Best Value of",tracepoint.quality,"found")
-            print("Time Elapsed: ",time.time()-start_time)
-        else:
-            # Timeout has occured break out of loop
-            break
-    p.join(0)
-    print("Algorithm Finished in",time.time()-start_time,"seconds")
-    p.terminate()
+            trace.add_tracepoint( tracepoint_read.recv() )
+
+    p.join( 0 ) # Join the subprocess which automatically closes the pipes
+    if p.is_alive():
+        p.terminate()
+
     return solution, trace
